test(governance): keep memory recall coverage within architecture budgets
